+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Add `--ping-interval` flag to override server-specified ping interval for improved job pickup performance [#3469](https://github.com/buildkite/agent/pull/3469) (@jasonwbarnett)
+
+### Changed
+- Change ping interval logging from debug to info level for better visibility [#3469](https://github.com/buildkite/agent/pull/3469) (@jasonwbarnett)
+
 ## [v3.108.0](https://github.com/buildkite/agent/tree/v3.108.0) (2025-10-02)
 [Full Changelog](https://github.com/buildkite/agent/compare/v3.107.2...v3.108.0)